Revert "fix: preserve collapsed effect rows on selection (#8659)"
This reverts commit ed58db2b91e6150c77ff31706027fb34697b9e1b.
